Changed NewPost and fixed multiline chat display
NewPost now fetches results when you ask for them, but has a 10 second
timeout after which it returns an error. It caches this data for up to 1
minute, but this can be overridden by specifying the post count as a
negative number instead of a positive one.

Also fixed is outgoing multi-line chats. It doesn't split them up
currently, but it does put each on it's own line with the normal chat
line header.